modify log cache and random functions

+++ src/test/sockproto.cpp
@@ -45,7 +45,7 @@ int GenSockProto::parseData(int fd,
     int ret = 0;
     
     while (0 < size) {
-        if (cache->m_pos < DEF_MSG_HEAD_SIZE) {
+        if ((int)cache->m_pos < DEF_MSG_HEAD_SIZE) {
             ret = parseHead(cache, buf, size);
         } else {
             ret = parseBody(fd, cache, buf, size);
@@ -90,7 +90,7 @@ int GenSockProto::parseHead(SockBuffer* buffer,
     bool bOk = false;
     char* psz = NULL;
     
-    if (buffer->m_pos < DEF_MSG_HEAD_SIZE) {
+    if ((int)buffer->m_pos < DEF_MSG_HEAD_SIZE) {
         cnt = DEF_MSG_HEAD_SIZE - buffer->m_pos;
         psz = &buffer->m_head[buffer->m_pos];
